I've moved all but the cython functions into their own files
