+++ pystoc/analytics/He3.py
@@ -1,0 +1,4 @@
+# This is the "probabilist's Hermite polynomial", which is different to the
+# "physicist's Hermite polynomials" used by SciPy
+def He3(y):
+    return y**3-3*y
