+++ pystoc/analytics/He4.py
@@ -1,0 +1,4 @@
+# This is the "probabilist's Hermite polynomial", which is different to the
+# "physicist's Hermite polynomials" used by SciPy
+def He4(y):
+    return y**4-6*y+3
