+++ pystoc/analytics/He6.py
@@ -1,0 +1,4 @@
+# This is the "probabilist's Hermite polynomial", which is different to the
+# "physicist's Hermite polynomials" used by SciPy
+def He6(y):
+    return y**6-15*y**4+45*y**2-15
