+++ pystoc/numerics/histogram_analytical_normalisation.py
@@ -1,0 +1,6 @@
+from numpy import diff
+
+
+# Returns the normalisation factor for a histogram, including one with weights
+def histogram_analytical_normalisation(bins, num_sims):
+    return num_sims*diff(bins)[0]
